// applications/webapp/src/app/api/generated/model/item-result.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemResult { 
    url?: string;
    status?: ItemResultStatusEnum;
    itemId?: string;
    reason?: string;
}
export enum ItemResultStatusEnum {
    Accepted = 'ACCEPTED',
    Rejected = 'REJECTED'
};