// applications/webapp/src/app/api/generated/model/run-summary.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RunSummary { 
    id?: string;
    status?: string;
    phase?: string;
    errorCode?: string;
    error?: string;
    videoUrl?: string;
    videoId?: string;
    channelName?: string;
    videoTitle?: string;
    videoCount?: number;
    createdAt?: string;
    updatedAt?: string;
}